updating primary processing workflow test
